requirements.txt
pandas
yfinance
python-dotenv
psycopg2
SQLAlchemy
